Move `user_has_perms_cache` into top-level `utils` module
