+++ src/model/channel/guild_channel.rs
@@ -189,7 +189,7 @@ impl GuildChannel {
         #[cfg(feature = "cache")]
         {
             if let Some(cache) = cache_http.cache() {
-                utils::user_has_perms_cache(
+                crate::utils::user_has_perms_cache(
                     cache,
                     self.id,
                     Some(self.guild_id),
@@ -310,7 +310,7 @@ impl GuildChannel {
         #[cfg(feature = "cache")]
         {
             if let Some(cache) = cache_http.cache() {
-                utils::user_has_perms_cache(
+                crate::utils::user_has_perms_cache(
                     cache,
                     self.id,
                     Some(self.guild_id),
@@ -414,7 +414,7 @@ impl GuildChannel {
         #[cfg(feature = "cache")]
         {
             if let Some(cache) = cache_http.cache() {
-                utils::user_has_perms_cache(
+                crate::utils::user_has_perms_cache(
                     cache,
                     self.id,
                     Some(self.guild_id),
@@ -965,7 +965,8 @@ impl GuildChannel {
             if let Some(cache) = cache_http.cache() {
                 let req = Permissions::SEND_MESSAGES;
 
-                if let Ok(false) = utils::user_has_perms(&cache, self.id, Some(self.guild_id), req)
+                if let Ok(false) =
+                    crate::utils::user_has_perms(&cache, self.id, Some(self.guild_id), req)
                 {
                     return Err(Error::Model(ModelError::InvalidPermissions(req)));
                 }
